+++ controllers/BaseController.php
@@ -10,7 +10,6 @@ namespace humhub\modules\wiki\controllers;
 use humhub\modules\content\components\ContentContainerController;
 use humhub\modules\wiki\models\WikiPage;
 use humhub\modules\wiki\permissions\AdministerPages;
-use humhub\modules\wiki\permissions\CreatePage;
 use humhub\modules\wiki\permissions\ViewHistory;
 
 
@@ -32,7 +31,7 @@ abstract class BaseController extends ContentContainerController
      */
     public function canCreatePage()
     {
-        return $this->contentContainer->permissionManager->can(CreatePage::class);
+        return (new WikiPage($this->contentContainer))->content->canEdit();
     }
 
     /**